requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
discord.py==2.5.2
frozenlist==1.6.0
gTTS==2.5.4
idna==3.10
multidict==6.4.4
propcache==0.3.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
requests==2.32.3
typing_extensions==4.13.2
urllib3==2.4.0
watchdog==6.0.0
yarl==1.20.0
aiohttp
